move cmdserver into main binary
