+++ cmd/direktiv/cmdserver/pkg/server/logging_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/direktiv/direktiv/cmd/cmd-exec/pkg/server"
+	"github.com/direktiv/direktiv/cmd/dirkeitv/cmdserver/pkg/server"
 	"github.com/stretchr/testify/assert"
 )
